refactor(storage): split `checksum` module
This will make it easier to expose only the main trait as public.
