+++ src/storage/src/storage/perform_upload.rs
@@ -15,7 +15,10 @@
 use super::client::{StorageInner, apply_customer_supplied_encryption_headers};
 use crate::model::Object;
 use crate::retry_policy::ContinueOn308;
-use crate::storage::checksum::{ChecksumEngine, ChecksummedSource, Known};
+use crate::storage::checksum::{
+    ChecksumEngine,
+    details::{ChecksummedSource, Known},
+};
 use crate::storage::client::info::X_GOOG_API_CLIENT_HEADER;
 use crate::storage::v1;
 use crate::upload_source::{IterSource, Seek, StreamingSource};
